llama-flask-app/requirements.txt
Flask==2.0.3
transformers==4.21.1
python-dotenv==0.19.2
torch==1.12.1